+++ src/main/java/gov/ornl/rse/datastreams/ssm_bats_rest_api/services/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * SSM BATS REST API services package information.
+ */
+package gov.ornl.rse.datastreams.ssm_bats_rest_api.services;
